+++ vlib/vweb/tests/vweb_test_server.v
@@ -80,6 +80,7 @@ pub fn (mut app App) user_repo_settings(username string, repository string) vweb
 [post]
 ['/json_echo']
 pub fn (mut app App) json_echo() vweb.Result {
+	eprintln('>>>>> received http request at /json_echo is: $app.req')
 	app.set_content_type(app.req.headers['Content-Type'])
 	return app.ok(app.req.data)
 }
@@ -87,6 +88,7 @@ pub fn (mut app App) json_echo() vweb.Result {
 // Make sure [post] works without the path
 [post]
 pub fn (mut app App) json() vweb.Result {
+	eprintln('>>>>> received http request at /json is: $app.req')
 	app.set_content_type(app.req.headers['Content-Type'])
 	return app.ok(app.req.data)
 }